baseline_model/requirements.txt
django
numpy
opencv-python
ultralytics
pandas
openpyxl
Pillow
pytesseract
